Issue #1909 - remove deprecated Unsafe usage
