+++ Google/guice-core/src/main/java/com/google/inject/internal/aop/HiddenClassDefiner.java
@@ -18,10 +18,6 @@ package com.google.inject.internal.aop;
 
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodHandles.Lookup;
-import java.lang.reflect.Array;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-
 /**
  * {@link ClassDefiner} that defines classes using {@code MethodHandles.Lookup#defineHiddenClass}.
  *
@@ -31,63 +27,7 @@ final class HiddenClassDefiner implements ClassDefiner {
 
   @Override
   public Class<?> define(Class<?> hostClass, byte[] bytecode) throws Exception {
-    Module guiceModule = HiddenClassDefiner.class.getModule();
-    Module hostModule = hostClass.getModule();
-    if (guiceModule.isNamed() && hostModule.isNamed()) {
-      if (!guiceModule.canRead(hostModule)) {
-        guiceModule.addReads(hostModule);
-      }
-      if (!hostModule.isOpen(hostClass.getPackageName(), guiceModule)) {
-        hostModule.addOpens(hostClass.getPackageName(), guiceModule);
-      }
+    Lookup hostLookup = MethodHandles.privateLookupIn(hostClass, MethodHandles.lookup());
+    return hostLookup.defineClass(bytecode);
     }
-    try {
-      Lookup lookup;
-      if (guiceModule.equals(hostModule)) {
-        lookup = MethodHandles.privateLookupIn(hostClass, MethodHandles.lookup());
-      } else {
-        lookup = MethodHandles.lookup().in(hostClass);
-      }
-      return defineHiddenClass(lookup, bytecode, hostClass);
-    } catch (Throwable t) {
-      throw t instanceof Exception ? (Exception) t : new RuntimeException(t);
-    }
-  }
-
-  private Class<?> defineHiddenClass(Lookup lookup, byte[] bytecode, Class<?> hostClass) throws Exception {
-    try {
-      return lookup.defineHiddenClass(bytecode, false, Lookup.ClassOption.NESTMATE).lookupClass();
-    } catch (IllegalAccessException e) {
-      // 1. Try getModuleLookup()
-      try {
-        Method getModuleLookup = hostClass.getDeclaredMethod("getModuleLookup");
-        getModuleLookup.setAccessible(true);
-        Lookup nextLookup = (Lookup) getModuleLookup.invoke(null);
-        if (nextLookup != null && !nextLookup.equals(lookup)) {
-          return nextLookup.defineHiddenClass(bytecode, false, Lookup.ClassOption.NESTMATE).lookupClass();
-        }
-      } catch (Throwable ignored) {
-      }
-
-      // 2. Try privateLookupIn
-      try {
-        Lookup nextLookup = MethodHandles.privateLookupIn(hostClass, MethodHandles.lookup());
-        if (nextLookup != null && !nextLookup.equals(lookup)) {
-          return nextLookup.defineHiddenClass(bytecode, false, Lookup.ClassOption.NESTMATE).lookupClass();
-        }
-      } catch (IllegalAccessException ignored) {
-      }
-
-      // 3. Try lookup().in()
-      Lookup fallbackLookup = MethodHandles.lookup().in(hostClass);
-      if (!fallbackLookup.equals(lookup)) {
-        try {
-          return fallbackLookup.defineHiddenClass(bytecode, false, Lookup.ClassOption.NESTMATE).lookupClass();
-        } catch (IllegalAccessException ignored) {
-        }
-      }
-
-      throw e;
-    }
-  }
 }
